Minor safety to ensure that the build happens correctly

+++ website/play/docker/build-initial-image.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/env bash
 
+if [ $(dirname "$0") != '.' ]; then
+    echo "Must call this with ./build-initial-image.sh"
+    exit 1
+fi
+
 set -euo pipefail
 cp -r ../../../template .
 rm -rf template/target
